requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
coverage==7.5.1
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.0.3
fonttools==4.51.0
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.7
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
nose==1.3.7
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2==2.9.10
py==1.11.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
randomize==0.14
redis==5.0.4
requests==2.31.0
rich==14.0.0
rich-toolkit==0.14.6
selenium==4.21.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
spotipy==2.23.0
SQLAlchemy==2.0.41
starlette==0.46.2
trio==0.25.1
trio-websocket==0.11.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.2.1
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.21.0
